requirements.txt
langchain==0.2.11
python-dotenv==1.1.1
langchain-community==0.2.10
pypdf2==3.0.1
groq
langchain_groq